Revert "Combine volcano & kube batch e2e tests"
